Housekeeping: merge two separate test modules for common.utils into one, for easier discoverability
